Move VizMovingHead to its own package.
